requirements.txt
asgiref==3.8.1
Django==5.1.3
pip==23.2.1
setuptools==68.2.2
sqlparse==0.4.4
typing-extensions==4.8.0
wheel==0.41.2
boto3>=1.35.0
ipython==8.16.1
django-storages>=1.14.4
celery==5.3.4
redis[hiredis]==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.7.0
pydantic<2.0
requests==2.31.0
structlog==23.2.0
stream-unzip==0.0.88
pytz==2023.3.post1
django-cors-headers==4.3.0
djangorestframework==3.14.0
PyJWT==2.8.0
beautifulsoup4==4.12.2
html5lib==1.1
Pillow
Pygments==2.16.1
gunicorn==21.2.0
uvicorn==0.23.2
django-prometheus==2.3.1
django-extensions==3.2.3
psycopg==3.0.1
psycopg2-binary

legacy-cgi
